fix: always overwrite branch even if PR were already closed
Fixes #884

Before:
When a PR was closed but the PR's branch (e.g. `npm-audit-fix-action/fix`) was not deleted,
pushing a new branch (but the same name) fails.

After:
Even if a PR was closed, this change always tries overwriting the PR's branch.
If a branch is not created yet, creating a branch passes as before.